Added educational program label to UI and moved class to another package
