+++ src/main/java/com/pnu/dev/pnufeedback/processor/FeedbackSubmissionProcessor.java
@@ -1,4 +1,4 @@
-package com.pnu.dev.pnufeedback.service;
+package com.pnu.dev.pnufeedback.processor;
 
 import com.pnu.dev.pnufeedback.dto.FeedbackSubmissionDto;
